Verify input source type constructors (#3779)
* Verify input source type constructors

* rename

* fix help link

* Documentation for GQL010

* docs

* format

* Update BOM test with detail

* Debug CallerFilePath

* Fix bom tests

---------
