+++ docs2/site/docs/analyzers/.vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "[markdown]": {
         "editor.rulers": [80],
-        "editor.wordWrap": "off"
+        "editor.wordWrap": "bounded"
     }
 }
