+++ src/GraphQL.Analyzers/Helpers/DiagnosticIds.cs
@@ -11,4 +11,5 @@ public static class DiagnosticIds
     public const string CAN_NOT_SET_SOURCE_FIELD = "GQL007";
     public const string DO_NOT_USE_OBSOLETE_ARGUMENT_METHOD = "GQL008";
     public const string USE_ASYNC_RESOLVER = "GQL009";
+    public const string CAN_NOT_RESOLVE_INPUT_SOURCE_TYPE_CONSTRUCTOR = "GQL010";
 }
