+++ src/GraphQL.ApiTests/BomTests.cs
@@ -10,10 +10,12 @@ public class BomTests
     {
         string GetPath([CallerFilePath] string path = "") => path; // <GIT_ROOT>\src\GraphQL.ApiTests\BomTests.cs
 
-        var gitRoot = new DirectoryInfo(GetPath()).Parent!.Parent!.Parent!;
+        var callerFilePath = GetPath();
+        var gitRoot = new FileInfo(callerFilePath).Directory!.Parent!.Parent!;
         // Protection from situations when this test is copied to another repo or the folder structure changed, etc.
-        if (!File.Exists(Path.Combine(gitRoot.FullName, "src", "GraphQL.sln")))
-            throw new InvalidOperationException("Unable to find repository root");
+        var slnFile = Path.Combine(gitRoot.FullName, "src", "GraphQL.sln");
+        if (!File.Exists(slnFile))
+            throw new InvalidOperationException($"Unable to find repository root - '{slnFile}' missing; caller file path '{callerFilePath}'.");
 
         byte[] buffer = new byte[3];
         int counter = 0;
